Fix clamp op decomposition
In  clamp op, `min` and `max` values are not casted to input data type correctly. Therefore, `max` is defaulted to `float32` whereas the input is `min`. This caused phi 3.5 vision to fail with the following error: `RuntimeError: Expected tensor for argument #1 'indices' to have one of the following scalar types: Long, Int; but got torch.FloatTensor instead (while checking arguments for embedding)`

+++ tt_torch/dynamo/decompositions.py
@@ -290,10 +290,10 @@ def clamp(
         max = float("inf")
 
     if not isinstance(max, torch.Tensor):
-        max = torch.ones(1, dtype=input.dtype) * max
+        max = (torch.ones(1, dtype=input.dtype) * max).to(input.dtype)
 
     if not isinstance(min, torch.Tensor):
-        min = torch.ones(1, dtype=input.dtype) * min
+        min = (torch.ones(1, dtype=input.dtype) * min).to(input.dtype)
 
     out = input
     if max.numel() <= out.numel():
